scroll left fix, more mana formatting
