+++ src/components/Teaser.jsx
@@ -5,7 +5,7 @@ export const Teaser = () => {
 
     return (
         <section id={"teaser"}>
-            <YouTube className={'teaser__video'} videoId={'fj5Te3rOxf0'} opts={{width : '100%', height: '100%'}}/>
+            <YouTube className={'teaser__video'} videoId={'fj5Te3rOxf0'} opts={{width : '100%', controls : 0, height: '100%'}}/>
             <div className={"teaser__poster"}></div>
         </section>
     )
