+++ src/components/events/Events.jsx
@@ -51,12 +51,18 @@ export const Events = () => {
 
     const scrollRef = useRef();
 
+    const handleScrollTo = direction => {
+        let left = direction === "left" ?  -scrollRef.current.parentNode.offsetWidth / 2 : scrollRef.current.parentNode.offsetWidth / 2;
+        left += scrollRef.current.scrollLeft;
+        scrollRef.current.scrollTo({left, behavior : "smooth"})
+    }
+
     return (
         <section id={'events'}>
             <div className={'events__title'}>
 
                 {isOverflowing && (
-                    <button name={'scroll-left'} onClick={()=>scrollRef.current.scrollTo({left : -scrollRef.current.parentNode.offsetWidth / 2, behaviour : "smooth"})}>
+                    <button name={'scroll-left'} onClick={()=>handleScrollTo("left")}>
                         <svg width="14" height="26" viewBox="0 0 14 26" fill="none" xmlns="http://www.w3.org/2000/svg">
                             <path d="M13 25L1 13L13 1" stroke="white" strokeWidth="2" strokeLinejoin="round"/>
                         </svg>
@@ -66,7 +72,7 @@ export const Events = () => {
                 <h2>{title}</h2>
 
                 {isOverflowing && (
-                    <button name={'scroll-right'} onClick={()=>scrollRef.current.scrollTo({left : scrollRef.current.parentNode.offsetWidth / 2, behaviour : "smooth"})}>
+                    <button name={'scroll-right'} onClick={()=>handleScrollTo("right")}>
                         <svg width="14" height="26" viewBox="0 0 14 26" fill="none" xmlns="http://www.w3.org/2000/svg">
                             <path d="M1 1L13 13L1 25" stroke="white" strokeWidth="2" strokeLinejoin="round"/>
                         </svg>
