+++ src/components/marketplace/MarketPlace.jsx
@@ -25,7 +25,7 @@ export const MarketPlace = () => {
             <div className={'marketplace__title'}>
 
                 {isOverflowing && (
-                    <button name={'scroll-left'}  onClick={()=>scrollRef.current.scrollTo({left : scrollRef.current.parentNode.offsetWidth / 2, behaviour : "smooth"})}>
+                    <button name={'scroll-left'}  onClick={()=>scrollRef.current.scrollTo({left : scrollRef.current.parentNode.offsetWidth / 2, behavior : "smooth"})}>
                         <svg width="14" height="26" viewBox="0 0 14 26" fill="none" xmlns="http://www.w3.org/2000/svg">
                             <path d="M13 25L1 13L13 1" stroke="white" strokeWidth="2" strokeLinejoin="round"/>
                         </svg>
@@ -35,7 +35,7 @@ export const MarketPlace = () => {
                 <h2>{content?.data?.content?.marketplace?.title}</h2>
 
                 {isOverflowing && (
-                    <button name={'scroll-right'}  onClick={()=>scrollRef.current.scrollTo({left : scrollRef.current.parentNode.offsetWidth / 2, behaviour : "smooth"})}>
+                    <button name={'scroll-right'}  onClick={()=>scrollRef.current.scrollTo({left : scrollRef.current.parentNode.offsetWidth / 2, behavior : "smooth"})}>
                         <svg width="14" height="26" viewBox="0 0 14 26" fill="none" xmlns="http://www.w3.org/2000/svg">
                             <path d="M1 1L13 13L1 25" stroke="white" strokeWidth="2" strokeLinejoin="round"/>
                         </svg>
